# discard performance improvements