--- a/crates/but-workspace/src/tree_manipulation/mod.rs
+++ b/crates/but-workspace/src/tree_manipulation/mod.rs
@@ -13,6 +13,5 @@ pub(super) mod discard_worktree_changes;
 pub(super) mod move_between_commits;
 pub(super) mod remove_changes_from_commit_in_stack;
 
-mod file;
 pub(crate) mod hunk;
 mod utils;