When adding a big folder with untracked files, like `node_modules`, then presently the `create_wd_tree` function runs twice, adding objects for a large number of files and leaving a possibly huge amount of loose objects.

This PR fixes this by bringing back the more surgical approach of directly editing the affected files on disk.

Another side-effect of this approach is that once again it finds the passed changes in the actual worktree-changes to assure they (still) match before getting to work.

The downside of this change is that once again there are known shortcomings around when submodules are checked out as part of a discard (that restores them).

Another side-effect is that discarding all hunks in a changed file that also has its executable bit changed will require another discard to undo the executable bit change. This is considered beneficial as it leaves a choice to the user.
Also, conflicts once again can't be discarded as it's unclear how to do that without an addition to the information that discard receives. At least this was now clearly described as a TODO, as it's certainly very useful to be able to discard conflicts.

**Interestingly, this test found another performance issue due to slow rename tracking when asking for worktree changes,
which needs a fix in `gitoxide`.** This would affect the entire experience as worktree-changes is called a lot (despite the UI currently not being able to show this many untracked files anyway).

### Performance Research

The testsetup:

```
small ( gitbutler/workspace) +13 -1 [$!?] via 
❯ gst
On branch gitbutler/workspace
Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
        modified:   hello

Untracked files:
  (use "git add <file>..." to include in what will be committed)
        node_modules/

no changes added to commit (use "git add" and/or "git commit -a")
```

```
small ( gitbutler/workspace) +13 -1 [$!?] via 
❯ find node_modules -type f  | wc -l
   69538
```

`node_modules` was copied from the GitButler repository.

#### Previously with `create_wd_tree`

![Screenshot 2025-05-19 at 07 46 14](https://github.com/user-attachments/assets/f3181240-2e99-4922-9eef-3d24d944aa29)

10s total runtime.

#### "surgical discard" with worktree-changes rename tracking bug

![Screenshot 2025-05-19 at 10 40 32](https://github.com/user-attachments/assets/e67bf89c-60d2-48b0-9f1a-eb567ada097e)

21s runtime, 20s of which are rename tracking.

#### "surgical discard" - fixed gix rename tracking

![Screenshot 2025-05-19 at 11 39 23](https://github.com/user-attachments/assets/7c105b5d-3607-411f-9889-b264a0923898)

4s total runtime, with about ~200ms for the main thread operation and dealing  for the status listing 70k untracked files (a problem on its own).

### Tasks

* [x] bring back previous implementation to discard worktree changes
* [x] adjust tests
* [x] fix `worktree-changes` performance issue
* [x] #8628

### Notes for the Reviewer

* The previous implementation was brutally copied, adapted, and dumped into `discard_worktree_changes.rs` to be self-contained.
* Nothing else was touched, unless it was required to fix compiler warnings, typically of now unused items.